1209b15

+++ LogonDlg.cpp
@@ -154,7 +154,7 @@ void SaveRunAsPassword(DWORD SessionId,LPTSTR RunAsUser,LPTSTR UserName,LPTSTR P
     return;
   DATA_BLOB pw={0};
   DATA_BLOB entropy={sizeof(KEYPASS),KEYPASS};
-  DATA_BLOB PW={(_tcslen(Password)+1)*sizeof(TCHAR),(BYTE*)Password};
+  DATA_BLOB PW={(DWORD)(_tcslen(Password)+1)*sizeof(TCHAR),(BYTE*)Password};
   if (!CryptProtectData(&PW,0,&entropy,0,0,CRYPTPROTECT_UI_FORBIDDEN,&pw))
     DBGTrace1("CryptProtectData failed: %s",GetLastErrorNameStatic());
   if (pw.cbData)
